chore: Extract function to estimate DLC channel fee reserve
